Fix bugs and organize blueprints
